% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_commit.R
\name{authors_commit}
\alias{authors_commit}
\title{Get Authors of a Commit}
\usage{
authors_commit(file_name)
}
\arguments{
\item{file_name}{a file in the git repository}
}
\value{
a data frame with the total amount of commits per user
}
\description{
Get Authors of a Commit
}
\examples{
\dontrun{
authors_commit(file_name = "README.md")
}
}
\author{
Debora Camacho
}
