% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file}
\alias{check_file}
\title{Check File Existence}
\usage{
check_file(file_name)
}
\arguments{
\item{file_name}{a file in the git repository}
}
\value{
A character phrase if the file does not exist and if empty
}
\description{
Check File Existence
}
\examples{
check_file(file_name = "check.R")
}
\author{
Debora Camacho
}
