% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Table Summary}
\usage{
create_summary_table(file_name, repo_path = getwd())
}
\arguments{
\item{file_name}{a file in the git repository}

\item{repo_path}{retrieves the current working directory}
}
\value{
a list of commit ids associated with each user
}
\description{
Table Summary
}
\examples{
\dontrun{
create_summary_table(file_name = "README.md")
}
}
\author{
Debora Camacho, Nicole Sanchez Flores, Ananda Wilkinson
}
