% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunk_name.R
\name{extract_name}
\alias{extract_name}
\title{Extracting Hunk Names}
\usage{
extract_name(file_name)
}
\arguments{
\item{file_name}{A string containing the path to a file tracked in a Git repository.}
}
\value{
A character vector of unique usernames from the file's commit history.
}
\description{
Extracting Hunk Names
}
\examples{
extract_name(file_name = "README.md")
}
\author{
Nicole Flores Sanchez
}
