% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_lines.R
\name{get_all_commit_shas}
\alias{get_all_commit_shas}
\title{Commit Changes}
\usage{
get_all_commit_shas(...)
}
\arguments{
\item{repo_path}{Provides the path to the repository, in this case it's our current working directory}

\item{all_commits}{Provides a list of all the commits made in the file}

\item{sha_author_list}{Provides a list of all the authors in the project}

\item{sha}{Represents the hash code for each commit made}

\item{message}{Represents the description of the changes made by the author}
}
\value{
A summary table with the commit hashes (SHA and their corresponding authors and message.
}
\description{
This function allows users to find commits for each group member in their project. It helps users identify each members contribution to the project, making it easier to determine who contributed the most, the least, or if they was an equal contribution.
}
\examples{
get_all_commit_shas()
}
\author{
Ananda Wilkinson
}
