% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_hunk.R
\name{hunks}
\alias{hunks}
\title{Getting Hunks from a File}
\usage{
hunks(file_name)
}
\arguments{
\item{file_name}{a file in the git repository}
}
\value{
A list of hunks, the commit history
}
\description{
Getting Hunks from a File
}
\examples{
\dontrun{
hunks(file_name = "README.md")
}
}
\author{
Nicole Sanchez Flores
}
