% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_readme}
\alias{install_readme}
\title{Install a private directory to store chchchanges README}
\usage{
install_readme(dir = ".")
}
\arguments{
\item{dir}{root directory of the git repository}
}
\description{
Install a private directory to store chchchanges README
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_readme()
\dontshow{\}) # examplesIf}
}
