% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{summarize_git_with_llm}
\alias{summarize_git_with_llm}
\alias{has_gemini_key}
\alias{init_chat}
\alias{is_chat_ready}
\title{Use an LLM to summarize git contributions}
\usage{
summarize_git_with_llm(
  repo = ".",
  chat = init_chat(),
  show_prompts = FALSE,
  user_prompt = NULL,
  ...
)

has_gemini_key()

init_chat(...)

is_chat_ready(chat)
}
\arguments{
\item{repo}{path to git repository}

\item{chat}{An \link[ellmer:Chat]{ellmer::Chat}}

\item{show_prompts}{Output the prompts as messages? Default is \code{FALSE}}

\item{user_prompt}{Additional text to add to the prompt}

\item{...}{arguments passed to \link[ellmer:Chat]{ellmer::Chat}}
}
\description{
Use an LLM to summarize git contributions
}
\examples{
\dontshow{if (has_gemini_key() && git2r::in_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
summarize_git_with_llm(show_prompts = TRUE)
summarize_git_with_llm(show_prompts = TRUE, user_prompt = "Respond only in Spanish.")

\dontshow{\}) # examplesIf}
has_gemini_key()

\dontshow{if (has_gemini_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
init_chat()
\dontshow{\}) # examplesIf}
\dontshow{if (has_gemini_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
chat <- init_chat()
is_chat_ready(chat)
\dontshow{\}) # examplesIf}
}
