% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2r.R, R/summarize.R
\name{tbl_blame}
\alias{tbl_blame}
\alias{summarize_blame}
\title{Summarize the blame for a file}
\usage{
tbl_blame(...)

summarize_blame(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[git2r:blame]{git2r::blame()}}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with the total amount of commits per user
}
\description{
Summarize the blame for a file
}
\examples{
\dontshow{if (git2r::in_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
tbl_blame(path = "README.md")
}
\dontshow{\}) # examplesIf}
\dontshow{if (git2r::in_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
summarize_blame(path = "README.md")
}
\dontshow{\}) # examplesIf}
}
\author{
Debora Camacho
}
