% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2r.R, R/summarize.R
\name{tbl_commits}
\alias{tbl_commits}
\alias{summarize_commits}
\title{Summarize the commit history}
\usage{
tbl_commits(...)

summarize_commits(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[git2r:commits]{git2r::commits()}}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with aggregated commit statistics for each user
}
\description{
Summarize the commit history

Summarize various git artifacts
}
\examples{
\dontshow{if (git2r::in_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tbl_commits()
\dontshow{\}) # examplesIf}
\dontshow{if (git2r::in_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
summarize_commits()
}
\dontshow{\}) # examplesIf}
}
\author{
Debora Camacho, Nicole Sanchez Flores, Ananda Wilkinson
}
