% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunk_name.R
\name{extract_name}
\alias{extract_name}
\title{Extracting Hunk Names}
\usage{
extract_name(file_name)
}
\arguments{
\item{hunk}{a list of the commit history from a specific file}
}
\value{
a list of the usernames will appear
}
\description{
Extracting Hunk Names
}
\examples{
extract_name(hunk = "hunk_name.R")
}
\author{
Nicole Flores Sanchez
}
