% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_lines.R
\name{get_all_commit_shas}
\alias{get_all_commit_shas}
\title{Commit Changes}
\usage{
get_all_commit_shas(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[git2r:repository]{git2r::repository()}}}
}
\value{
A summary table with the commit hashes (SHA and their corresponding authors and message.
}
\description{
This function allows users to find commits for each group member in their project. It helps users identify each members contribution to the project, making it easier to determine who contributed the most, the least, or if they was an equal contribution.
}
\examples{
\dontrun{
get_all_commit_shas()
}
}
\author{
Ananda Wilkinson
}
