% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{llm_summarize_git}
\alias{llm_summarize_git}
\title{Use an LLM to summarize git contributions}
\usage{
llm_summarize_git(repo = ".", ...)
}
\arguments{
\item{repo}{path to git repository}

\item{...}{currently ignored}
}
\description{
Use an LLM to summarize git contributions
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
llm_summarize_git()

\dontshow{\}) # examplesIf}
}
