% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{summarize_git_with_llm}
\alias{summarize_git_with_llm}
\alias{init_chat}
\alias{is_chat_ready}
\title{Use an LLM to summarize git contributions}
\usage{
summarize_git_with_llm(
  repo = ".",
  chat = init_chat(),
  show_prompts = FALSE,
  user_prompt = NULL,
  ...
)

init_chat(...)

is_chat_ready(chat)
}
\arguments{
\item{repo}{path to git repository}

\item{show_prompts}{Output the prompts as messages? Default is \code{FALSE}}

\item{user_prompt}{Additional text to add to the prompt}

\item{...}{arguments passed to \code{\link[ellmer:chat_gemini]{ellmer::chat_gemini()}}}
}
\description{
Use an LLM to summarize git contributions
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
summarize_git_with_llm(show_prompts = TRUE)
summarize_git_with_llm(show_prompts = TRUE, user_prompt = "Respond only in Spanish.")

\dontshow{\}) # examplesIf}
init_chat()

chat <- init_chat()
is_chat_ready(chat)

}
